Added an in statement targeting actions authroised by a superuser with an error message if user does not have authorization to manage products

+++ products/views.py
@@ -74,6 +74,10 @@ def individual_product(request, product_id):
 @login_required
 def add_product(request):
     """ Add a product to the store """
+    if not request.user.is_superuser:
+        messages.error(request, 'You do not have store authorization to carry out this request')
+        return redirect(reverse('home'))
+
     if request.method == 'POST':
         form = ProductForm(request.POST, request.FILES)
         if form.is_valid():
@@ -96,6 +100,10 @@ def add_product(request):
 @login_required
 def update_product(request, product_id):
     """ Edit an entry """
+    if not request.user.is_superuser:
+        messages.error(request, 'You do not have store authorization to carry out this request')
+        return redirect(reverse('home'))
+
     product = get_object_or_404(Product, pk=product_id)
     if request.method == 'POST':
         form = ProductForm(request.POST, request.FILES, instance=product)
@@ -122,6 +130,10 @@ def update_product(request, product_id):
 @login_required
 def delete_product(request, product_id):
     """ Deleting a product """
+    if not request.user.is_superuser:
+        messages.error(request, 'You do not have store authorization to carry out this request')
+        return redirect(reverse('home'))
+
     product = get_object_or_404(Product, pk=product_id)
     product.delete()
     messages.success(request, f'{product.name} has been deleted.')
